bringing the two together (#17)
* Accidently pulled from the wrong branch (#16)

* This should work, fixed ItemRedCrown

* Update en_US.lang

* Infusion matrix (#6)

* Create zh_CN.lang

* Update en_US.lang

* Some interceptor code cleanup

Private fields, pruned unnecessary try/catches, simplified logic, some variables renamed for clarity and consistency

* Move all essentia in a single tick if it exists

* spotlessApply

* Update BS

* Fixing all the stuff for the soap

* Update README.md

* Update README.md


* Infusion matrix (#7)

* Create zh_CN.lang

* Update en_US.lang

* Some interceptor code cleanup

Private fields, pruned unnecessary try/catches, simplified logic, some variables renamed for clarity and consistency

* Move all essentia in a single tick if it exists

* spotlessApply

* Update BS

* Update README.md

* Update README.md


* Fixing a small glitch

* This should fix the error I was having


* Bring the two together (#17)

* Update README.md

* update BS+SA

* Update README.md

* Fix server crash with renderer (#18)
