+++ src/main/java/thaumicinsurgence/main/ThaumicInsurgence.java
@@ -1,8 +1,5 @@
 package thaumicinsurgence.main;
 
-import cpw.mods.fml.client.registry.ClientRegistry;
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 import cpw.mods.fml.common.Mod;
 import cpw.mods.fml.common.SidedProxy;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
@@ -10,18 +7,8 @@ import cpw.mods.fml.common.event.FMLMissingMappingsEvent;
 import cpw.mods.fml.common.event.FMLMissingMappingsEvent.MissingMapping;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
-import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
 import thaumicinsurgence.main.utils.LogHelper;
 import thaumicinsurgence.main.utils.VersionInfo;
-import thaumicinsurgence.renderers.blockrenderers.BlockAlphaPedestalRenderer;
-import thaumicinsurgence.renderers.blockrenderers.BlockMatrixAlphaRenderer;
-import thaumicinsurgence.renderers.blockrenderers.BlockPillarAlphaRenderer;
-import thaumicinsurgence.renderers.tileentityrenderers.TileAlphaPedestalRenderer;
-import thaumicinsurgence.renderers.tileentityrenderers.TileMatrixAlphaRenderer;
-import thaumicinsurgence.renderers.tileentityrenderers.TilePillarAlphaRenderer;
-import thaumicinsurgence.tileentity.TileEntityInfusionMatrixAlpha;
-import thaumicinsurgence.tileentity.TileEntityInfusionPillarAlpha;
-import thaumicinsurgence.tileentity.TileEntityPedestalAlpha;
 
 @SuppressWarnings("unused")
 @Mod(
@@ -50,31 +37,6 @@ public class ThaumicInsurgence {
     @Mod.EventHandler
     public void postInit(FMLPostInitializationEvent event) {
         proxy.postInit(event);
-        specialRenderers();
-    }
-
-    public void registerTileEntitySpecialRenderer(Class tile, TileEntitySpecialRenderer renderer) {
-        ClientRegistry.bindTileEntitySpecialRenderer(tile, renderer);
-    }
-
-    public void registerBlockRenderer(ISimpleBlockRenderingHandler renderer) {
-        RenderingRegistry.registerBlockHandler(renderer);
-    }
-
-    public void specialRenderers() {
-
-        this.registerTileEntitySpecialRenderer(TileEntityInfusionMatrixAlpha.class, new TileMatrixAlphaRenderer(0));
-        Config.blockStoneDeviceRI = RenderingRegistry.getNextAvailableRenderId();
-        this.registerBlockRenderer(new BlockMatrixAlphaRenderer());
-
-        this.registerTileEntitySpecialRenderer(TileEntityInfusionPillarAlpha.class, new TilePillarAlphaRenderer());
-        Config.blockStoneDeviceTwoRI = RenderingRegistry.getNextAvailableRenderId();
-        this.registerBlockRenderer(new BlockPillarAlphaRenderer());
-
-        // TODO: FIX THIS
-        Config.blockStoneDeviceThreeRI = RenderingRegistry.getNextAvailableRenderId();
-        this.registerTileEntitySpecialRenderer(TileEntityPedestalAlpha.class, new TileAlphaPedestalRenderer());
-        this.registerBlockRenderer(new BlockAlphaPedestalRenderer());
     }
 
     @Mod.EventHandler
